+++ client/Ntreev.Crema.Client.Converters/MenuItems/ExportRevisionTableMenuItem.cs
@@ -27,6 +27,7 @@ using System.ComponentModel.Composition;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Ntreev.Crema.Spreadsheet.Excel;
 
 namespace Ntreev.Crema.Client.Converters.MenuItems
 {
