+++ client/Ntreev.Crema.Client.Converters/Spreadsheet/ViewModels/ImporterViewModel.cs
@@ -37,6 +37,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.WindowsAPICodePack.Dialogs;
 using System.ComponentModel;
+using Ntreev.Crema.Spreadsheet.Excel;
 using Ntreev.Library.Linq;
 using Ntreev.ModernUI.Framework.ViewModels;
