+++ common/Ntreev.Crema.Commands.Spreadsheet.Sharing/SpreadSheetCommand.cs
@@ -38,6 +38,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics;
+using Ntreev.Crema.Spreadsheet.Excel;
+using Ntreev.Crema.Spreadsheet.Excel;
 
 namespace Ntreev.Crema.Commands.Spreadsheet
 {
