+++ common/Ntreev.Crema.Spreadsheet/Excel/SpreadsheetWriterSettings.cs
@@ -16,24 +16,42 @@
 //OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Ntreev.Library;
 
-namespace Ntreev.Crema.Spreadsheet
+namespace Ntreev.Crema.Spreadsheet.Excel
 {
-    public static class SpreadsheetUtility
+    public class SpreadsheetWriterSettings
     {
-        public static string Ellipsis(string name)
-        {
-            if (name.Length <= 31)
-                return name;
+        public static SpreadsheetWriterSettings Default = new SpreadsheetWriterSettings();
 
-            var middle = "~" + name.GetHashCode() + "~";
-            var first = name.Substring(0, (name.Length - middle.Length) / 2);
-            var last = name.Substring(name.Length - (31 - (first.Length + middle.Length)));
-            return first + middle + last;
+        public SpreadsheetWriterSettings()
+        {
+            this.OmitSignatureDate = true;
+            this.Tags = TagInfo.All;
+            this.NameEllipsis = SpreadsheetUtility.Ellipsis;
+        }
+
+        public bool OmitAttribute
+        {
+            get;
+            set;
+        }
+
+        public bool OmitSignatureDate
+        {
+            get;
+            set;
+        }
+
+        public TagInfo Tags
+        {
+            get;
+            set;
+        }
+
+        public Func<string, string> NameEllipsis
+        {
+            get; set;
         }
     }
 }
