+++ common/Ntreev.Crema.Spreadsheet/Ntreev.Crema.Spreadsheet.csproj
@@ -64,11 +64,11 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <Compile Include="SpreadsheetUtility.cs" />
-    <Compile Include="SpreadsheetWriter.cs" />
-    <Compile Include="SpreadsheetReader.cs" />
+    <Compile Include="Excel\SpreadsheetUtility.cs" />
+    <Compile Include="Excel\SpreadsheetWriter.cs" />
+    <Compile Include="Excel\SpreadsheetReader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SpreadsheetWriterSettings.cs" />
+    <Compile Include="Excel\SpreadsheetWriterSettings.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\sub-modules\Ntreev.Library\Ntreev.Library\Ntreev.Library.csproj">
@@ -89,6 +89,7 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="..\Ntreev.Crema.AssemblyInfo\Ntreev.Crema.AssemblyInfo.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
