+++ tools/Ntreev.Crema.Comparer/Tables/ViewModels/TableBrowserViewModel.cs
@@ -27,6 +27,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
+using Ntreev.Crema.Spreadsheet.Excel;
 
 namespace Ntreev.Crema.Comparer.Tables.ViewModels
 {
